refactor: Move from stores to context.
